Fix LP64 str() overload clash + vcpkg binary caching
- tools.{h,cpp}: str(int64_t) -> str(long long). On LP64 Linux int64_t IS long, so
  str(int64_t) was a redefinition of str(long); long long is the original MinGW
  __int64 and is distinct from long everywhere. Verified with act (GCC 13.3, Ubuntu).
- release.yml: cache vcpkg-built packages in the GitHub Actions cache.

+++ source/tools.cpp
@@ -263,10 +263,10 @@ std::string str(uint32_t value)
 	return buf;
 }
 
-std::string str(int64_t value)
+std::string str(long long value)
 {
 	char buf[128];
-	snprintf(buf, sizeof(buf), "%lld", (long long)value);
+	snprintf(buf, sizeof(buf), "%lld", value);
 	return buf;
 }
